ruleguard: suggest t.Setenv in test cases
Where an existing testcase has a pattern of:
```
os.Setenv("KEY", "VALUE")
defer os.Unsetenv("KEY")
```

Suggest replacement of:
```
t.Setenv("KEY")
```

As this will handle the `defer` for you and will restore the previous
value, rather than unsetting it. It will also ensure a compile time
error is raised if the test has been marked as `t.Parallel` (which would
be unsafe).